% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds_find.R
\name{ds_ls}
\alias{ds_ls}
\title{ds_ls}
\usage{
ds_ls(proj = "somWASH")
}
\value{
named list
}
\description{
helper function to return list of all datasets used  in project. Includes short/abbreviated name and file path
}
\examples{
\dontrun{
library(somaliaWASH)
ds_ls())
}

}
