% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2r_hex_aggregate.R
\name{h2r_hex_aggregate_pct}
\alias{h2r_hex_aggregate_pct}
\title{h2r_hex_aggregate_pct}
\usage{
h2r_hex_aggregate_pct(df, hex, summarise_cols, country_code = "som")
}
\arguments{
\item{df}{settlement level h2r data set}

\item{hex}{hex sf object}

\item{summarise_cols}{cols to summarise}
}
\value{

}
\description{
h2r_hex_aggregate_pct
}
\examples{
\dontrun{
h2r_hex_aggregate_pct(settlement_level,
                      som_hex,
                      summarise_cols = "lack_food_reasons.noland",
                     country_code = "som")

}
}
