% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2r_hex_aggregate.R
\name{h2r_to_sf_utm}
\alias{h2r_to_sf_utm}
\title{h2r_to_sf_utm}
\usage{
h2r_to_sf_utm(df, country_code = "som")
}
\arguments{
\item{country_code}{country_code (default= "som")}

\item{settlement}{level H2R data.frame with coordinates stored as character "geometry" column}
}
\value{
sf object projected to UTM zone based on country code
}
\description{
helper function used inside h2r_aggregate funcs
}
\examples{
\dontrun{
settlement_level |>
  h2r_to_sf_utm
}
}
