% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2r_hex_aggregate.R
\name{hex_aggregate_pt_count}
\alias{hex_aggregate_pt_count}
\title{Title}
\usage{
hex_aggregate_pt_count(
  pt_sf = swp_sf,
  hex = som_hex,
  var = NULL,
  var_val = NULL,
  country_code = "som"
)
}
\arguments{
\item{pt_sf}{sf class point object}

\item{hex}{hex/grid/polygon}

\item{var}{variable of interest (default=NULL) . If NULL, just point counts of entire dataset are aggregated}

\item{var_val}{variable value of interest (default = NULL). If NULL, just point counts of entire dataset are aggregated}

\item{country_code}{three chr country code (default = "som")}
}
\value{

}
\description{
Title
}
\examples{
\dontrun{
library(somaliaWASH)
swp <-  load_swp()
swp_sf <- swp |>
    sf::st_as_sf(coords=c("longitude","latitude"), crs=4326)

# just aggregate point counts
water_point_count_per_grid <- swp_sf |>
   hex_aggregate_pt_count(hex = som_hex)

# aggregate point counts based on `var`- `val` combination
strategic_water_point_count_per_grid <- swp_sf |>
   hex_aggregate_pt_count(hex = som_hex,var = "is_strategic",var_val = "yes")
}
}
