% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_h2r_data.R
\name{load_h2r_data}
\alias{load_h2r_data}
\title{load_h2r_data
Provide the location of the main folder in our example we were reading from OneDrive, the the path to the csv file.}
\usage{
load_h2r_data(
  main_folder_path = Sys.getenv("SOM_2203_WASH_DATA"),
  data_path = "data/202209/raw/REACH_H2R_0522-2607_data2.csv"
)
}
\value{

}
\description{
function to load H2R data set - will use environment key to specify base data directory
}
