% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_h2r_indictors.R
\name{load_h2r_indicators}
\alias{load_h2r_indicators}
\title{load_h2r_indictors
Calling the loading and selecting functions to prepare the data.}
\usage{
load_h2r_indicators()
}
\value{

}
\description{
combination of `load_h2r_data` and `select_h2r_indicators` to quicky load only data with only relelvant indicators
}
