% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_swp.R
\name{load_swp}
\alias{load_swp}
\title{load strategic water point data}
\usage{
load_swp()
}
\value{
data.frame containing strategic water point data
}
\description{
load strategic water point data
}
\examples{
\dontrun{
swp <-  load_swp()
}
}
