% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_h2r_indicators.R
\name{select_h2r_indicators}
\alias{select_h2r_indicators}
\title{select_h2r_indicators
This function is for selecting appropriate indicators for a given country. You need to supply data frame and the country code like "som" for Somalia}
\usage{
select_h2r_indicators(input_df, country_code = "som")
}
\value{

}
\description{
function to select relevant h2r indicators from data set
}
